Remove and rename index files, and remove redundant folders
rename files
