+++ tavla/app/(admin)/components/Navbar.tsx
@@ -7,7 +7,7 @@ import Image from 'next/image'
 import Link from 'next/link'
 import { usePathname } from 'next/navigation'
 import { useFeatureFlagEnabled } from 'posthog-js/react'
-import { Login } from './Login'
+import { Login } from './Login/Login'
 import { MobileNavbar } from './MobileNavbar'
 
 function Navbar({ loggedIn }: { loggedIn: boolean }) {
