+++ tavla/app/(admin)/mapper/[id]/page.tsx
@@ -2,8 +2,8 @@ import { ButtonGroup } from '@entur/button'
 import { FolderIcon } from '@entur/icons'
 import { Heading1, Label, Paragraph } from '@entur/typography'
 import { getBoardsForFolder } from 'app/(admin)/actions'
-import { CreateBoard } from 'app/(admin)/components/CreateBoard'
-import { DeleteFolder } from 'app/(admin)/components/Delete'
+import { CreateBoard } from 'app/(admin)/components/CreateBoard/CreateBoard'
+import { DeleteFolder } from 'app/(admin)/components/DeleteFolder/DeleteFolder'
 import { BoardTable } from 'app/(admin)/oversikt/components/BoardTable'
 import EmptyOverview from 'app/(admin)/oversikt/components/EmptyOverview'
 import { BreadcrumbsNav } from 'app/(admin)/tavler/[id]/BreadcrumbsNav'
@@ -15,8 +15,8 @@ import { notFound, redirect } from 'next/navigation'
 import { getFolder } from 'src/firebase'
 import type { FolderDB } from 'src/types/db-types/folders'
 import type { UserDB } from 'src/types/db-types/users'
-import { MemberAdministration } from '../components/MemberAdministration'
-import { UploadLogo } from '../components/UploadLogo'
+import { MemberAdministration } from '../components/MemberAdministration/MemberAdministration'
+import { UploadLogo } from '../components/UploadLogo/UploadLogo'
 
 export type TProps = {
     params: Promise<{ id: FolderDB['id'] }>
