+++ tavla/app/(admin)/oversikt/page.tsx
@@ -8,8 +8,8 @@ import { initializeAdminApp } from 'app/(admin)/utils/firebase'
 import { getUserFromSessionCookie } from 'app/(admin)/utils/server'
 import type { Metadata } from 'next'
 import { redirect } from 'next/navigation'
-import { CreateBoard } from '../components/CreateBoard'
-import { CreateFolder } from '../components/CreateFolder'
+import { CreateBoard } from '../components/CreateBoard/CreateBoard'
+import { CreateFolder } from '../components/CreateFolder/CreateFolder'
 import { DemoBoardBanner } from './components/DemoBoardBanner'
 import { FoldersAndBoardsContent } from './components/FoldersAndBoardsContent'
