+++ tavla/app/(admin)/tavler/[id]/rediger/components/DuplicateBoard/DuplicateBoard.tsx
