+++ tavla/app/(admin)/tavler/[id]/rediger/components/Preview.tsx
