+++ tavla/app/(admin)/tavler/[id]/rediger/components/Settings/Settings.tsx
