+++ tavla/app/(admin)/tavler/[id]/rediger/components/TileCard/TileCard.tsx
