+++ tavla/app/components/WordCarousel.tsx
